+++ PokemonEncyclopedia/PokemonEncyclopedia/FeedCache/FeedStore.swift
@@ -1,0 +1,16 @@
+//
+//  FeedStore.swift
+//  PokemonEncyclopedia
+//
+//  Created by Ankit on 07/07/21.
+//
+
+import Foundation
+
+public protocol FeedStore {
+	typealias DeletionCompletion = (Error?)->Void
+	typealias InsertionCompletion = (Error?)->Void
+
+	func deleteCachedFeed(completion: @escaping DeletionCompletion)
+	func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
