+++ PokemonEncyclopedia/PokemonEncyclopedia/FeedCache/LocalFeedLoader.swift
@@ -36,10 +36,3 @@ public class LocalFeedLoader {
 	}
 }
 
-public protocol FeedStore {
-	typealias DeletionCompletion = (Error?)->Void
-	typealias InsertionCompletion = (Error?)->Void
-
-	func deleteCachedFeed(completion: @escaping DeletionCompletion)
-	func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
-}
